requirements.txt
discord.py
Flask
python-dotenv
requests
